Lesson 1: Implement fourth attempt, replacing mult/div in loop with Bresenham 'error' technique.
Use derror as the steps for increasing the 'error', which is a 0.5 threshold for
deciding whether or not to increase y by 1.

The algorithm increases the error by derror for each step of x (each loop iteration).
It'll add 1 to y when error > 0.5, in which case it adds 1 to y and also subtracts
1 from error.